requirements.txt
torch
matplotlib
tqdm
PyYAML
python-box==6.0.2
ensure==1.0.2
fastapi==0.78.0
uvicorn==0.18.3
Jinja2==3.1.2
-e .